Fully buffer small RC files
